% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{agedata}
\alias{agedata}
\alias{agedata<-}
\title{Extract and assign agedata of Proxyzoo}
\usage{
agedata(x)

agedata(x) <- value
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Extract and assign agedata of Proxyzoo
}
