% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{agesumm}
\alias{agesumm}
\alias{agesumm<-}
\title{Extract and assign age summary of Proxyzoo}
\usage{
agesumm(x)

agesumm(x) <- value
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Extract and assign age summary of Proxyzoo
}
