% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytibble.R
\name{apply_proxy}
\alias{apply_proxy}
\title{Apply a function to all proxy data in a Proxytibble}
\usage{
apply_proxy(x, fun, ...)
}
\arguments{
\item{x}{`PTBoxProxydata::Proxytibble`}

\item{fun}{function}

\item{...}{arguments passed on to `fun`}

\item{col}{column containing proxy data (in `zoo::zoo` or `PTBoxProxydata::Proxyzoo` format) to apply `fun` to}
}
\value{

}
\description{
Apply a function to all proxy data in a Proxytibble
}
