% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytibble.R
\name{as_Proxytibble}
\alias{as_Proxytibble}
\title{Conversion of `data.frame`s and `list` of `matrix` objects into `Proxytibble`.}
\usage{
as_Proxytibble(x, ...)
}
\arguments{
\item{x}{either proxydata (`list` of `matrix` objects) or of class `data.frame`/`tibble::tibble` holding an entire set of paleo data, including agedata and metadata. See \emph{Details} for the expected format.}

\item{agedata}{agedata, if provided, it has to be of same class as `x`. Ignored if `x` is of `class` `data.frame/tibble::tibble`. See \emph{Details} for the expected format.}

\item{agesumm}{summary of the agedata, if provided, it has to be of same class as `x`. Ignored if `x` is of `class` `data.frame/tibble::tibble`. See \emph{Details} for the expected format.}

\item{metadata}{meta data corresponding to each entity. If `x` is a `list` of proxy data, `metadata` has to be of class `data.frame/tibble::tibble`. Ignored if `x` is of class `data.frame/tibble::tibble`.}

\item{zoo_format}{format of the object containing the proxy data, can be 'zoo' or 'Proxyzoo'. See `vignette('zoo-support-sdm')` for details.}
}
\value{

}
\description{
Conversion of `data.frame`s and `list` of `matrix` objects into `Proxytibble`.
}
\section{Details}{

For input format `data.frame`/`tibble::tibble` the `data.frame`/`tibble::tibble` is expected to hold all metadata as plain rows and corresponding proxy data, age data, and age summary as nested `data.frame`/`tibble::tibble`. See \emph{Examples}.
The column names holding proxy, age, and age summary data can be specified in the package conbfiguration. Per default, they are `proxydata`, `agedata`, `agesummdata`.
Mandatory metadata entries are (in default configuration of `PTBoxProxydata`): INSERT
If agesummary is provided, these columns are mandatory: mean_age, q25, q50, q75, giving the mean and the 25th, 50th, and 75th quantile of the age model ensemble. If not provided, they are computed based on `agedata`. Additinal `agesumm` statistics, like error ranges, can be provided.

For input format `matrix`, `x` is expected to be a list of `matrix` objects which hold the proxy data as columns and correspond to the `metadata`. Analogous formats are expected for `agedata` and `agesumm` if provided. The foremat of `metadata` is expected to be a named list, with list names corresponding to the metadata names (as above) and vector entries holding the metadata that corresponds to the data passed as matrix lists. See \emph{Examples}.
}

