% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{as_Proxyzoo}
\alias{as_Proxyzoo}
\title{Convert one (two) zoo::zoo objects to one PTBoxProxydata::Proxyzoo}
\usage{
as_Proxyzoo(x, ...)
}
\arguments{
\item{x}{`zoo::zoo` object}

\item{...}{passed to `as_Proxyzoo.zoo`}
}
\value{

}
\description{
Convert one (two) zoo::zoo objects to one PTBoxProxydata::Proxyzoo
}
