% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{as_Proxyzoo.zoo}
\alias{as_Proxyzoo.zoo}
\title{Convert one (two) zoo::zoo objects to one PTBoxProxydata::Proxyzoo}
\usage{
\method{as_Proxyzoo}{zoo}(proxyzoo, agezoo = NULL, indexcol = "index")
}
\arguments{
\item{proxyzoo}{`zoo::zoo`, contains proxy data (see *Details* for expected formatting)}

\item{agezoo}{`zoo::zoo`, optional contains age/age model data (see *Details* for expected formatting)}

\item{indexcol}{}
}
\value{
object of `PTBoxProxydata::Proxyzoo` (see *Details* for return formatting, \link{Proxyzoo} for shape of `PTBoxProxydata::Proxyzoo`)
}
\description{
Convert one (two) zoo::zoo objects to one PTBoxProxydata::Proxyzoo
}
\details{
Expects proxy data `proxyzoo` either with a column named `depth` or `age` which can be used as an index 
(unless names are specified otherwise in the configuration of the `PTBoxProxydata` package), or tries to use the first column
of the `proxyzoo` `zoo::zoo` object. If `zoo::zoo` object `agezoo` is supplied additionally, expects 
matching `depth`, `age` or first indexing column.

If both `depth` and `age` are contained in `proxyzoo`, `depth` will be used 
as *unambiguous* index, and `age` will be included in `Proxyzoo$agedata`. 

If only `age` is provided as index,
this *can* (does not have to) lead to *ambigious* indexing of the data in case that a hiatus is present in the data.

If none of `depth` or `age` are provided, returned `Proxyzoo$agedata` will only consist of an `$index`/`index()` column
linking to `Proxyzoo$agedata` (see \link{Proxyzoo} for shape of `PTBoxProxydata::Proxyzoo`).
}
