% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{as_zoo.data.frame}
\alias{as_zoo.data.frame}
\title{Conversion to zoo::zoo respecting PTBox conventions}
\usage{
\method{as_zoo}{data.frame}(proxyframe, ...)
}
\arguments{
\item{proxyframe}{}

\item{...}{ignored}
}
\value{

}
\description{
Conversion to zoo::zoo respecting PTBox conventions
}
