% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{as_zoo}
\alias{as_zoo}
\alias{as_data.frame}
\title{Type conversion of Proxyzoo}
\usage{
as_zoo(x, ...)

as_data.frame(x, ...)
}
\arguments{
\item{x}{}
}
\value{



}
\description{
Type conversion of Proxyzoo
}
