% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyDataManager.R
\name{cache_RDS}
\alias{cache_RDS}
\title{Cache dataset(s) from a ProxyDataManager to RDS}
\usage{
cache_RDS(x, ...)
}
\arguments{
\item{mng}{ProxyDataManager}

\item{dataset_names}{character, single or vector of datasets to load or 'all' to cache all available datasets.}

\item{output_format}{character, one of c('Proxytibble', 'tibble'). 'tibble' corresponds to `tibble::tibble`}

\item{zoo_format}{character, one of c('zoo', 'zoo_STACY'). format of the proxy data. See \emph{Details} and `vignette('zoo-support-sdm')`.}

\item{RDS_root}{character, root directory for pre-formatted .RDS data. See \emph{Details} for more.}

\item{force_file}{logical, if TRUE forces to load all `dataset_names` from file and to recreate the corresponding RDS caches.}
}
\value{
dataset(s) in specified format
}
\description{
Cache dataset(s) from a ProxyDataManager to RDS
}
\section{Details}{

Datasets are always cached to RDS and meta data entries are written if they are loaded for the first time. This can be triggered when
supplying the respective `dataset_names` to \code{\link{load_set}} or to \code{\link{cache_RDS}}.

If directory specified as \code{RDS_root} does not exist, will try to create it simply (i.e. non-recursively). 
Throws an error if \code{RDS_root} cannot be created in this way. In this case check the path you submitted and
potentially create the necessary directories. This functionality is kept rudimentary by intention to avoid 
filling storage spaces unintentionally.

\code{zoo_format} is ignored if \code{output_format == 'tibble'} because proxy data as `zoo::zoo`/`PTBoxProxydata::zoo_STACY` objects is only 
supported by `PTBox` inside the `PTBoxProxydata::Proxytibble` conventions.
}

