% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{depth}
\alias{depth}
\alias{depth<-}
\title{Extract and assign depth of Proxyzoo}
\usage{
depth(x)

depth(x) <- value
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Extract and assign depth of Proxyzoo
}
