% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyDataManager.R
\name{export_master_sheet}
\alias{export_master_sheet}
\title{Export the master sheet to file}
\usage{
export_master_sheet(x, ...)
}
\arguments{
\item{mng}{`ProxyDataManager` object}

\item{path}{file path for output file}

\item{format}{output format one of `yml` (default), `csv`, `xlsx`}

\item{overwrite}{logical, overwrite destination `path` if already exists? (default `FALSE`)}
}
\value{
None
}
\description{
Export the master sheet to file
}
\examples{
\dontrun{
mng <- ProxyDataManager()
export_master_sheet(mng, 'path/to/output_file.format_output_file', format = 'csv')
}
}
