% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytibble.R
\name{is.Proxytibble}
\alias{is.Proxytibble}
\title{Class check for `PTBoxProxydata::Proxytibble`}
\usage{
is.Proxytibble(x)
}
\arguments{
\item{x}{an object of class 'Proxytibble'}
}
\description{
Class check for `PTBoxProxydata::Proxytibble`
}
