% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyDataManager.R
\name{load_set}
\alias{load_set}
\title{Load dataset(s) from a ProxyDataManager}
\usage{
load_set(x, ...)
}
\arguments{
\item{mng}{ProxyDataManager}

\item{dataset_names}{character, single or vector of datasets to load. See \emph{Details}}

\item{proxy_names}{character, single or vector of proxies to load. See \emph{Details}.}

\item{output_format}{character, one of c('Proxytibble', 'tibble'). 'tibble' corresponds to `tibble::tibble`}

\item{zoo_format}{character, one of c('zoo', 'zoo_STACY'). format of the proxy data. See \emph{Details} and `vignette('zoo-support-sdm')`.}

\item{only_mandatory_attr}{logical, defaults `FALSE`, only return the mandatory meta data in the output}

\item{RDS_root}{character, root directory for pre-formatted .RDS data. See \emph{Details} for more.}

\item{force_file}{logical, if TRUE forces to load all `dataset_names` from file and to recreate the corresponding RDS caches.}
}
\value{
dataset(s) in specified format
}
\description{
Load dataset(s) from a ProxyDataManager
}
\section{Details}{

If loading by `dataset_names` only, datasets will be loaded either from file or from cache and returned in format `output_format`. If 
`proxy_names` are additionally supplied, the datasets `dataset_names` are filtered for respective proxies and the filtered data will
be returned. This is only possible if datasets have previously been cached to RDS. If only `proxy_names` is specified all available 
datasets which contain those proxies will be filtered and returned. Again, this works only on datasets which have previously been 
cached to RDS.

Datasets are always cached to RDS and meta data entries are written if they are loaded for the first time. This can be triggered when
supplying the respective `dataset_names` to \code{\link{load_set}} or to \code{\link{cache_RDS}}.

If directory specified as \code{RDS_root} does not exist, will try to create it simply (i.e. non-recursively). 
Throws an error if \code{RDS_root} cannot be created in this way. In this case check the path you submitted and
potentially create the necessary directories. This functionality is kept rudimentary by intention to avoid 
filling storage spaces unintentionally.

\code{zoo_format} is ignored if \code{output_format == 'tibble'} because proxy data as `zoo::zoo`/`PTBoxProxydata::zoo_STACY` objects is only 
supported by `PTBox` inside the `PTBoxProxydata::Proxytibble` conventions.
}

\examples{
\dontrun{
# create a ProxyDataManager instance
mng <- ProxyDataManager()
# load the icecore_data with default options
icecore_data <- load_set(mng, 
                         dataset_names = 'icecore_data')
}
}
