% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_load_default.R
\name{manager_load_default}
\alias{manager_load_default}
\title{Load a preformatted dataset for the ProxyDataManager}
\usage{
manager_load_default(file, dataset_name, dataset_id, man_attr)
}
\arguments{
\item{file}{character, file containing the dataset}

\item{man_attr}{character, mandatory attributes for the ProxyDataManager}
}
\value{

}
\description{
Load a preformatted dataset for the ProxyDataManager
}
