% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_load_default.R
\name{manager_load_icecore_testset}
\alias{manager_load_icecore_testset}
\title{Load the test data set with EPICA DML and DC ice cores for the ProxyDataManager}
\usage{
manager_load_icecore_testset(file, dataset_name, dataset_id, zoo_format)
}
\arguments{
\item{file}{character, file containing the dataset}

\item{dataset_name}{character, (short) name of the dataset (passed on to helpers)}

\item{dataset_id}{numeric, id of the dataset (passed on to helpers)}

\item{zoo_format}{character, format of zoo to be used for storing proxy data (passed on to helpers)}
}
\value{

}
\description{
Load the test data set with EPICA DML and DC ice cores for the ProxyDataManager
}
\references{
The water isotope data from the EPICA ice cores on Dronning Maud Land (DML) and at Dome C (DC) are taken from
B. Stenni et al.: The deuterium excess records of EPICA Dome C and Dronning Maud Land ice cores (East Antarctica), Quat. Sci. Rev., 29 (1-2), 146-159, https://doi.org/10.1016/j.quascirev.2009.10.009, 2010
}
