% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_load_default.R
\name{manager_load_monticchio_testset}
\alias{manager_load_monticchio_testset}
\title{Load the test data set with arboreal pollen data from Lake Monticchio for the ProxyDataManager}
\usage{
manager_load_monticchio_testset(file, dataset_name, dataset_id, zoo_format)
}
\arguments{
\item{file}{character, file containing the dataset}

\item{dataset_name}{character, (short) name of the dataset (passed on to helpers)}

\item{dataset_id}{numeric, id of the dataset (passed on to helpers)}

\item{zoo_format}{character, format of zoo to be used for storing proxy data (passed on to helpers)}
}
\value{

}
\description{
Load the test data set with arboreal pollen data from Lake Monticchio for the ProxyDataManager
}
\references{
The pollen data from the Monticchio archive was first published by
Wulf, S., Kraml, M., Brauer, A., et al.: Tephrochronology of the 100 ka lacustrine sediment record of Lago Grande di Monticchio (southern Italy), Quatern. Int., 122, 7–30, https://doi.org/10.1016/j.quaint.2004.01.028, 2004.
The univariate arboreal pollen data is decribed in
Adam, M., Weitzel, N., and Rehfeld, K.: Identifying Global-Scale Patterns of Vegetation Change During the Last Deglaciation From Paleoclimate Networks, Paleoceanogr. Paleoclimatol., 36, 12, https://doi.org/10.1029/2021PA004265, 2021.
It builds on the harmonized ACER database by 
Sánchez Goñi, M. F., Desprat, S., Daniau, A.-L., et al.: The ACER pollen and charcoal database: a global resource to document vegetation and fire response to abrupt climate changes during the last glacial period, Earth Syst. Sci. Data, 9, 679–695, https://doi.org/10.5194/essd-9-679-2017, 2017.
}
