% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyDataManager.R
\name{preview_all}
\alias{preview_all}
\title{Title}
\usage{
preview_all(x)
}
\arguments{
\item{x}{object of class `ProxyDataManager`}
}
\value{
empty
}
\description{
Title
}
\examples{
mng <- ProxyDataManager()
preview_all(mng)
}
