% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyDataManager.R
\name{print.ProxyDataManager}
\alias{print.ProxyDataManager}
\title{Printing methods for the `ProxyDataManager`}
\usage{
\method{print}{ProxyDataManager}(mng)
}
\arguments{
\item{mng}{`ProxyDataManager`}
}
\value{

}
\description{
Printing methods for the `ProxyDataManager`
}
\examples{
mng <- ProxyDataManager()
print(mng)
}
