% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{proxydata}
\alias{proxydata}
\alias{proxydata<-}
\title{Extract and assign proxydata of Proxyzoo}
\usage{
proxydata(x)

proxydata(x) <- value
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Extract and assign proxydata of Proxyzoo
}
