% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytibble.R
\name{Proxytibble}
\alias{Proxytibble}
\title{Proxytibble S3 class}
\usage{
Proxytibble(...)
}
\arguments{
\item{...}{passed to \code{tibble::tibble}}
}
\value{
object of classes c('Proxytibble', 'tbl', 'tbl_df', 'data.frame')
}
\description{
Proxytibble S3 class
}
