% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{Proxyzoo}
\alias{Proxyzoo}
\title{Proxyzoo S3 class for universal handling of proxy data, including age model ensembles}
\usage{
Proxyzoo(proxydata = NULL, agedata = NULL, agesumm = NULL, depth = NULL)
}
\arguments{
\item{...}{passed to \link{as_Proxyzoo.zoo}}
}
\value{
object of class `Proxyzoo`. Empty if no data supplied via `...`.
}
\description{
Proxyzoo S3 class for universal handling of proxy data, including age model ensembles
}
\details{
See `vignette('zoo-support-PTBoxProxydata')` for details.
}
