% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{Proxyzoo_proxydata}
\alias{Proxyzoo_proxydata}
\alias{Proxyzoo_agedata}
\alias{Proxyzoo_agesumm}
\alias{Proxyzoo_depth}
\title{Dynamic column names for the Proxyzoo}
\usage{
Proxyzoo_proxydata()

Proxyzoo_agedata()

Proxyzoo_agesumm()

Proxyzoo_depth()
}
\description{
Column names as specified in `PTBoxProxydata-config.yml`
and read from config in `hooks.R` on attachment/loading of the package.
}
\details{
\code{Proxyzoo_proxydata()} # default `proxydata`

\code{Proxyzoo_agedata()} # default `agedata`

\code{Proxyzoo_agesumm()} # default `agesumm`

\code{Proxyzoo_depth()} # default `depth`
}
