% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{package}
\name{PTBoxProxydata-package}
\alias{PTBoxProxydata}
\alias{PTBoxProxydata-package}
\title{PTBoxProxydata: The Proxy Data Manager for a consistent, efficient and reusable data workflow for (paleo) time series within the PaleoToolBox (PTBox)}
\description{
This package provides the ProxyDataManager class and imposes conventions on format and processing of datasets used within the PaleoToolBox (PTBox) for a more consistent, efficient, and reusable data workflow.
}
\author{
\strong{Maintainer}: Moritz Adam \email{madam@iup.uni-heidelberg.de}

Authors:
\itemize{
  \item Beatrice Ellerhoff \email{beatrice.ellerhoff@iup.uni-heidelberg.de}
  \item Jean-Philippe Baudouin \email{Jean-Phillipe.Baudouin@mnf.uni-tuebingen.de}
  \item Nils Weitzel \email{nweitzel@iup.uni-heidelberg.de}
}

}
\keyword{internal}
