% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{PTBoxProxydata_print_config}
\alias{PTBoxProxydata_print_config}
\alias{PTBoxProxydata_get_config}
\alias{PTBoxProxydata_print_config_path}
\title{Print/get configuration of the PTBoxProxydata package and the corresponding configuration path}
\usage{
PTBoxProxydata_print_config()

PTBoxProxydata_get_config()

PTBoxProxydata_print_config_path()
}
\description{
Print/get configuration of the PTBoxProxydata package and the corresponding configuration path
}
\examples{
PTBoxProxydata_print_config_path()

# these are equivalent
cfg <- PTBoxProxydata_get_config()
print(cfg)
PTBoxProxydata_print_config()
}
