% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{PTBoxProxydata_reload_config}
\alias{PTBoxProxydata_reload_config}
\title{Reload the configuration for the PTBoxProxydata package}
\usage{
PTBoxProxydata_reload_config(
  config_path = get("CONFIG_PATH", .PTBoxProxydata),
  config = Sys.info()[["nodename"]]
)
}
\arguments{
\item{config_path}{path of the changed new or old `.yml` file containing the configuration. See \emph{Warning}.}

\item{config}{configuration to load from the `.yml` file provided in \code{config_path}. See \emph{Details}.}
}
\value{
no return
}
\description{
Reload the configuration for the PTBoxProxydata package
}
\section{Warning}{

Be careful when changing the default configuration of \emph{PTBoxProxydata}. This may turn your package instance to be unable to read the Master Spreadsheet temporarily.
If you made changes to the default configuration file, \code{config_path} doesn't have to be supplied. However, it is recommended to test changes with a 
custom configuration file first before changing the configuration at it's native location. See \code{vignette('how-to-use-sdm')} as well.
}

\section{Details}{

If \code{config} is not provided or \code{config = 'default'}, will load the configuration tagged with `default` in the config file. 
If custom config is provided but not complete, will always expand config with `default` configuration from config file. Be careful when changing the `default`
configuration for your package instance (See \emph{Warning} and \code{vignette('PTBoxProxydata_howto')}).
}

\section{See also}{

\code{\link{PTBoxProxydata_print_config}, \link{PTBoxProxydata_print_config_path}}
}

\examples{
\dontrun{
# print current config path
PTBoxProxydata_print_config_path()
# set a config file
new_config_file <- 'my/new/config/path/PTBoxProxydata-config.yml'
PTBoxProxydata_reload_config(config_path = new_config_file,
                             config = 'my_machine')
}
}
