% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{select_proxies}
\alias{select_proxies}
\title{Select proxies by name}
\usage{
select_proxies(x, ...)
}
\arguments{
\item{x}{`zoo::zoo` or `PTBoxProxydata::zoo_STACY`}

\item{proxy_names}{character vector of proxy names}
}
\value{

}
\description{
Select proxies by name
}
