% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxytibble.R
\name{update_proxy_names}
\alias{update_proxy_names}
\title{Update the `proxy_name` and `proxy_unit` column of a `Proxytibble` to (re-)match the `proxy_data` column}
\usage{
update_proxy_names(x, ...)
}
\arguments{
\item{x}{`PTBoxProxydata::Proxytibble`}

\item{fun}{function}

\item{dcol}{data column containing proxy data (in `zoo::zoo` or `PTBoxProxydata::Proxyzoo` format) to apply `fun` to}

\item{ncol}{data column containing proxy name(s)}

\item{ucol}{data column containing proxy unit(s)}
}
\value{

}
\description{
Update the `proxy_name` and `proxy_unit` column of a `Proxytibble` to (re-)match the `proxy_data` column
}
