% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{zoo_apply}
\alias{zoo_apply}
\title{Apply a function to all proxy data contained in a zoo::zoo (univariate or multivariate)
while allowing for changing the dimension of the underlying zoo (this is a less strict and
more generalized version of zoo_applyfix)}
\usage{
zoo_apply(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Apply a function to all proxy data contained in a zoo::zoo (univariate or multivariate)
while allowing for changing the dimension of the underlying zoo (this is a less strict and
more generalized version of zoo_applyfix)
}
