% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{zoo_apply.zoo}
\alias{zoo_apply.zoo}
\title{Apply a function to multivariate zoo::zoo}
\usage{
\method{zoo_apply}{zoo}(x, fun, out_index = NULL, ...)
}
\arguments{
\item{x}{}

\item{fun}{Function operating on a univariate zoo}

\item{...}{}
}
\value{

}
\description{
Apply a function to multivariate zoo::zoo
}
