% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{zoo_applyfix}
\alias{zoo_applyfix}
\title{Apply a function to all proxy data contained in a zoo::zoo (univariate or multivariate)
while maintaining the dimension/length of the zoo}
\usage{
zoo_applyfix(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Apply a function to all proxy data contained in a zoo::zoo (univariate or multivariate)
while maintaining the dimension/length of the zoo
}
