% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proxyzoo.R
\name{zoo_applyfix.zoo}
\alias{zoo_applyfix.zoo}
\title{Apply a function to multivariate zoo::zoo}
\usage{
\method{zoo_applyfix}{zoo}(x, fun, ...)
}
\arguments{
\item{x}{}

\item{fun}{Function operating on a zoo}

\item{...}{}
}
\value{

}
\description{
Apply a function to multivariate zoo::zoo
}
